requirements copy.txt
# KBS 국제 뉴스 크롤링 프로젝트 필수 패키지

# 웹 크롤링
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# 동적 페이지 크롤링 (Selenium)
selenium==4.16.0
webdriver-manager==4.0.1

# 데이터 처리 및 엑셀 저장
pandas==2.2.0
openpyxl==3.1.2

# OpenAI API
openai==1.58.1

# Streamlit 웹 앱
streamlit==1.31.0

# 환경변수 관리
python-dotenv==1.0.0

# 데이터 시각화
plotly==5.18.0
matplotlib==3.8.2
wordcloud==1.9.3
